requirements.txt
streamlit
azure-cognitiveservices-speech
fpdf
python-dotenv